refactor: update import paths to use 'src' prefix for consistency across the backend
