+++ backend/src/repositories/repositories/construction_approval_repo/construction_approval_repo.py
@@ -1,8 +1,8 @@
 from typing import Optional, List
 from sqlalchemy.orm import Session
 
-from repositories.models.construction import ConstructionApproval
-from .construction_approval_repo_interface import IConstructionApprovalRepo
+from src.repositories.models.construction import ConstructionApproval
+from src.repositories.repositories.construction_approval_repo.construction_approval_repo_interface import IConstructionApprovalRepo
 
 
 class ConstructionApprovalRepo(IConstructionApprovalRepo):
