+++ backend/src/repositories/repositories/construction_phase_repo/construction_phase_repo.py
@@ -2,8 +2,8 @@ from typing import Optional, List
 from sqlalchemy.orm import Session
 from sqlalchemy import asc
 
-from repositories.models.construction import ConstructionPhase
-from .construction_phase_repo_interface import IConstructionPhaseRepo
+from src.repositories.models.construction import ConstructionPhase
+from src.repositories.repositories.construction_phase_repo.construction_phase_repo_interface import IConstructionPhaseRepo
 
 
 class ConstructionPhaseRepo(IConstructionPhaseRepo):
