+++ backend/src/repositories/repositories/construction_phase_repo/construction_phase_repo_interface.py
@@ -1,7 +1,7 @@
 from abc import ABC, abstractmethod
 from typing import Optional, List
 
-from repositories.models.construction import ConstructionPhase
+from src.repositories.models.construction import ConstructionPhase
 
 
 class IConstructionPhaseRepo(ABC):
