+++ backend/src/repositories/repositories/construction_progress_repo/construction_progress_repo_interface.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 from typing import Optional, List
 from datetime import datetime
 
-from repositories.models.construction import ConstructionProgress
+from src.repositories.models.construction import ConstructionProgress
 
 
 class IConstructionProgressRepo(ABC):
