+++ backend/src/repositories/repositories/construction_repo/construction_repo_interface.py
@@ -2,7 +2,7 @@ from abc import ABC, abstractmethod
 from typing import Optional, List
 from datetime import datetime
 
-from repositories.models.construction import Construction
+from src.repositories.models.construction import Construction
 
 
 class IConstructionRepo(ABC):
