+++ backend/src/repositories/repositories/construction_repo/construction_repo_mock.py
@@ -1,7 +1,7 @@
 from typing import Optional, List
 from datetime import datetime
 
-from repositories.models.construction import Construction
+from src.repositories.models.construction import Construction
 from .construction_repo_interface import IConstructionRepo
