+++ backend/src/repositories/repositories/user_repo/user_repo.py
@@ -1,7 +1,7 @@
 from typing import Optional, List
 from sqlalchemy.orm import Session
 
-from repositories.models.user import User
+from src.repositories.models.user import User
 from .user_repo_interface import IUserRepo
